requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
Werkzeug==2.2.3
zipp==3.15.0
certifi==2022.12.7
charset-normalizer==3.1.0
geographiclib==2.0
geopy==2.3.0
idna==3.4
numpy==1.24.2
Pillow==9.4.0
requests==2.28.2
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
typing_extensions==4.5.0
urllib3==1.26.15
